# fiat-payments-backend/src/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BrainSafes API",
    "description": "API completa para el ecosistema BrainSafes - Educación descentralizada con IA, NFTs y DeFi",
    "version": "1.0.0",
    "contact": {
      "name": "BrainSafes Team",
      "email": "support@brainsafes.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.brainsafes.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/ipfs/upload": {
      "post": {
        "summary": "Subir archivo a IPFS",
        "description": "Subir un archivo al sistema IPFS",
        "tags": ["IPFS"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "pin": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archivo subido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "hash": {
                          "type": "string"
                        },
                        "size": {
                          "type": "number"
                        },
                        "url": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
} 